+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.21.5](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.21.4...@coveo/quantic@2.21.5) (2023-04-26)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.21.4](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.21.3...@coveo/quantic@2.21.4) (2023-04-25)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.21.3](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.21.2...@coveo/quantic@2.21.3) (2023-04-24)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.21.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.21.1...@coveo/quantic@2.21.2) (2023-04-18)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.21.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.21.0...@coveo/quantic@2.21.1) (2023-04-14)
+
+### Bug Fixes
+
+- **quantic:** condition that detects if more child results are available in collection fixed ([#2818](https://github.com/coveo/ui-kit/issues/2818)) ([7e2d1a4](https://github.com/coveo/ui-kit/commit/7e2d1a47c35212966d31b7e9d74c3a1df37d293d))
+- **quantic:** transforming the numberOfFoldedResults property to a numeric value ([#2816](https://github.com/coveo/ui-kit/issues/2816)) ([982958c](https://github.com/coveo/ui-kit/commit/982958c1bd1b76eef0e3523f566e918b584b1359))
+
+# [2.21.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.20.3...@coveo/quantic@2.21.0) (2023-04-13)
+
+### Features
+
+- **quantic:** Creation of a quanticQuickviewContent component ([#2797](https://github.com/coveo/ui-kit/issues/2797)) ([932e110](https://github.com/coveo/ui-kit/commit/932e110b369516be147f7b37ea978efd9834ecc0))
+- **quantic:** modified quanticResultLink to support salesforceResult in salesforce console Tab ([#2804](https://github.com/coveo/ui-kit/issues/2804)) ([bac4c98](https://github.com/coveo/ui-kit/commit/bac4c9893e4cbb1bfcd978bfbfb9d736cd903340))
+
+## [2.20.3](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.20.2...@coveo/quantic@2.20.3) (2023-04-12)
+
+**Note:** Version bump only for package @coveo/quantic
+
 ## [2.20.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.20.1...@coveo/quantic@2.20.2) (2023-04-12)
 
 **Note:** Version bump only for package @coveo/quantic